Added ICamera for camera abstraction Added 2d camera support Added IResized hook for when the screen is resized Added Vector class that caches vector objects

+++ src/main/java/apex/gl/IGameApp.java
@@ -10,7 +10,7 @@ public interface IGameApp {
 
     void render(GameStore store);
 
-    void update(GameStore store, double delta);
+    void update(GameStore store, float delta);
 
     Window createWindow();
 }
